Update patch set 1
Patch Set 1:

the ota utility did work... however it did not activate a Sprint phone. tried 3 times. 1) straight out the box only rooting the phone. 2) going to sprints website and adding the phone to the line (as u normally would to activate) and then running the ota utility. 3) on my evo that isnt active (but has been and has all the carrier numbers and such) that i swotched to. it seems whatever number its actually calling isnt a legit number/code on sprints network. it just says as if i dialed a bad number even on an active phone. on a unactivated phone it just calls and says that i need to activate the phone before making calls (this is normal for a sprint phone that isnt activated)

this however may work on other cdma networks such as vzw and so on. just doesnt seem to work on sprints network. 

i do however have logs of both this patch doing the ota and sprints sense rom doing it and it working. i plan to look further into it and see if i can add to the ota utility sprint support. this would be pretty useful along with other things like prl updating and so on.

Patch-set: 1