Update patch set 1
Patch Set 1:

im not to sure on the two new apn protocols. for some reason since 2.3.4 we are stuck with the v2 apn introduced in the distro update.
anyhow i just did what appeared to be needed to correctly use the v2 apn, since i wasnt able to find a way back to a v1 apn. this fixes mms send a receive on sprint devices.

Patch-set: 1