Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

I went ahead and changed it to not show the button backlights at higher levels.  It wouldn't break autobrightness as that was only for the button backlights, which are only on or off (meaning they don't have levels).  The stock behavior for this phone is to always show the button backlights.

Patch-set: 2